Refactor navigation resources and menu
